Update Querying, Add User login. - fixes of build fails

+++ SourceCode/se/ucsc/hsptl/assignment/db/dto/AttendanceDTO.java
@@ -1,14 +1,14 @@
 package se.ucsc.hsptl.assignment.db.dto;
 
 import java.sql.Time;
-import java.util.Date;
+import java.sql.Date;
 
 /**
  * Created by Indika on 5/4/2017.
  */
 public class AttendanceDTO
 {
-  private int attendanceId;
+  private long attendanceId;
 
   private String employeeId;
